Splitting offers section from main page
